Add files via upload
This is a version of the exercise without Spring MVC. 
I was unsure of the expectations of what the exercise would be (whether you had the time to build and run the project on Tomcat). But for design/decision-making, I chose to create this plain Java project as the simpler version of the exercise.